requirements.txt
asyncio==3.4.3
bleak==0.20.1
contourpy==1.0.7
cycler==0.11.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pyparsing==3.0.9
scipy==1.10.1
six==1.16.0
tqdm==4.65.0
paho-mqtt==1.6.1
